refactor: create file for types
