+++ packages/trader/src/Modules/Trading/Components/Form/Purchase/cancel-deal-info.tsx
@@ -5,7 +5,7 @@ import { isDesktop, isMobile, getDecimalPlaces } from '@deriv/shared';
 import { localize } from '@deriv/translations';
 import { observer } from '@deriv/stores';
 import { useTraderStore } from 'Stores/useTraderStores';
-import { TProposalTypeInfo } from '../../../Containers/purchase';
+import { TProposalTypeInfo } from '../../../../../../types';
 
 const CancelDealInfo = observer(({ proposal_info }: { proposal_info: TProposalTypeInfo }) => {
     const { currency, has_cancellation } = useTraderStore();
