+++ packages/trader/src/Modules/Trading/Containers/purchase.tsx
@@ -6,21 +6,8 @@ import PurchaseFieldset from 'Modules/Trading/Components/Elements/purchase-field
 import { getContractTypePosition } from 'Constants/contract';
 import { useTraderStore } from 'Stores/useTraderStores';
 import { observer, useStore } from '@deriv/stores';
+import { TProposalTypeInfo } from '../../../../types';
 import ContractInfo from 'Modules/Trading/Components/Form/Purchase/contract-info.jsx';
-
-export type TProposalTypeInfo = {
-    has_error?: boolean;
-    id: string;
-    has_increased?: boolean;
-    message?: string;
-    cancellation?: {
-        ask_price: number;
-        date_expiry: number;
-    };
-    growth_rate?: number;
-    returns?: string;
-    stake: string;
-};
 
 const Purchase = observer(({ is_market_closed }: { is_market_closed: boolean }) => {
     const {
