+++ packages/trader/types.ts
@@ -1,0 +1,13 @@
+export type TProposalTypeInfo = {
+    has_error?: boolean;
+    id: string;
+    has_increased?: boolean;
+    message?: string;
+    cancellation?: {
+        ask_price: number;
+        date_expiry: number;
+    };
+    growth_rate?: number;
+    returns?: string;
+    stake: string;
+};
